feat(integrations): the switch decides, and admission asks it
admit stops reading the per-connection mode and the two boolean switches and
asks the installation posture instead. Every automatic trigger goes through one
answer; a manual run never consults it, because a person pressing the button on
a contact has already decided.

The old switches asked which WRITER a purchase followed — a human typing a
contact, a connector importing one — and the answer differed because a
connector's thousands of contacts each spent credits. An automatic run now buys
only what costs nothing, so that distinction stopped paying for itself.

settings_test.go holds the claim that one reader answers the posture: it parses
the package and fails when any file outside settings.go reaches for the entry.
Mutation-checked — planting a second reader fails it.

admit moves to its own file. runs.go was 7 lines over the 500 cap with the
change in it, and admission answers a different question from queueing: the
installation's standing policy rather than this one subject.
