+++ backend/gates/rbacgate_test.go
@@ -185,6 +185,7 @@ var ungatedEntryPoints = gatekit.Waive(map[string]string{ // #nosec G101 -- waiv
 	"internal/modules/identity:Get":                          "onboarding wizard state, SELF-scoped: onboardingActor resolves the authenticated human and the query is keyed on user_id, so no object grant applies to your own checkpoint",
 	"internal/modules/identity:Put":                          "the write half of the same self-scoped wizard state; onboardingActor is the gate and the row is keyed on the acting user",
 	"internal/modules/integrations:ExecuteSubmit":            "worker execution (compose/jobs_providerruns.go) under the system principal: advances a run that QueueRun already object-gated (integrations grant) AND row-gated (EnsureVisible on the person) when it was queued; the run id comes from the durable job the same transaction committed, no request path reaches it, and the egress lease inside it bounds what the call may touch",
+	"internal/modules/integrations:Get":                      "SettingsStore.Get reads one integrations setting through platform/settings.Get, which delegates to Store.Raw — and Raw takes the object grant PER SETTING against the object the entry declares (auth.Require(def.Object(), ActionRead)), which for AutomaticLookup is integrations. internal/platform/settings is one of THIS gate's roots, so that gate is judged here directly rather than assumed; re-gating in the wrapper would re-check the same read against the same object, the capture:Get shape exactly. Update beside it DOES gate, because its nil-patch branch never reaches the settings write",
 	"internal/modules/integrations:RunDueSweep":              "worker sweep (compose/jobs_providerruns.go) under the system principal: drains runs already admitted at queue time — polls, claim recoveries and expiries on rows the sweep's own predicates select; no record is returned and no human actor exists on this path",
 	"internal/modules/overlay:BlockAutoMap":                  "same: only usermapservice.go calls it, behind requireUserMapAdmin",
 	"internal/modules/finance:RecordSyncFailure":             "the sweep's own bookkeeping for the pass it just failed, on the same connector-principal path as SyncConnection; it writes no tenant record, only the connection's attempt stamp, status and error code — and the audit row that names the transition, from the same principal that stamped the connection's own captured_by",
