+++ backend/internal/modules/integrations/runs.go
@@ -204,78 +204,6 @@ func (s *Store) resolveProvider(ctx context.Context, named string) (string, erro
 	}
 }
 
-// ErrProviderAmbiguous reports that an automatic trigger found more than one
-// connected provider and refused to guess which one to spend on. It is a
-// configuration fault an operator must resolve, never a state to swallow.
-var ErrProviderAmbiguous = errors.New("integrations: more than one provider is connected, so an automatic trigger cannot choose which to spend on")
-
-// admittedConnection is the connection state a run freezes itself against.
-type admittedConnection struct {
-	id         string
-	version    int64
-	epoch      int64
-	mode       string
-	autoCreate bool
-	autoImport bool
-	categories []string
-	refreshAge *int
-	dailyLimit *int
-}
-
-// admit resolves the connection and refuses the triggers its policy does not
-// admit. A connection that is not connected cannot spend anything, and a
-// trigger the customer switched off must not spend on their behalf.
-func (s *Store) admit(ctx context.Context, tx pgx.Tx, name string, trigger provider.Trigger) (admittedConnection, error) {
-	var c admittedConnection
-	err := tx.QueryRow(ctx, `
-		SELECT id::text, version, execution_epoch, mode, automatic_individual_create,
-		       automatic_import, categories, refresh_after_days, daily_run_limit
-		  FROM provider_connection
-		 WHERE provider = $1 AND status = 'connected'
-		 FOR SHARE`, name).
-		Scan(&c.id, &c.version, &c.epoch, &c.mode, &c.autoCreate,
-			&c.autoImport, &c.categories, &c.refreshAge, &c.dailyLimit)
-	if errors.Is(err, pgx.ErrNoRows) {
-		return admittedConnection{}, provider.ErrNotConnected
-	}
-	if err != nil {
-		return admittedConnection{}, fmt.Errorf("integrations: reading the connection: %w", err)
-	}
-
-	switch trigger {
-	case provider.TriggerAutomaticCreate:
-		if c.mode != "automatic_on_create" || !c.autoCreate {
-			return admittedConnection{}, errTriggerNotAdmitted
-		}
-	case provider.TriggerAutomaticImport:
-		if !c.autoImport {
-			return admittedConnection{}, errTriggerNotAdmitted
-		}
-	}
-	return c, nil
-}
-
-// errTriggerNotAdmitted reports that the saved policy does not run this
-// trigger. It is not an error the caller shows anybody: the person.created
-// consumer swallows it, because "auto-enrich is off" is the configuration
-// working, not a failure.
-var errTriggerNotAdmitted = errors.New("integrations: this trigger is not admitted by the saved policy")
-
-// IsTriggerNotAdmitted reports whether a QueueRun refusal was the saved
-// policy declining this trigger — auto-enrich switched off, or a mode that
-// does not run on create.
-//
-// It exists because the event consumer lives in compose and must tell that
-// refusal apart from a failure, and the alternative is comparing error text:
-// a predicate here means the sentinel can be reworded without silently
-// turning a swallowed configuration state into a logged error.
-func IsTriggerNotAdmitted(err error) bool {
-	// ErrNothingFreeToBuy is the same kind of answer: the saved policy leaves
-	// an automatic run nothing it may spend on, which is a configuration
-	// working rather than a failure to log.
-	return errors.Is(err, errTriggerNotAdmitted) || errors.Is(err, ErrNothingFreeToBuy)
-}
-
 // queueOne is the admission pipeline for one subject. Each refusal writes a
 // skipped run rather than nothing at all: a customer asking "why was this
 // person not enriched" deserves a row that answers, and a silent no-op cannot.
